refactor: 대시보드 라우터 개선
