--- src/pages/dashboard/components/CalendarAndEventList/index.ts
@@ -1,3 +1,0 @@
-import CalendarAndEventList from './CalendarAndEventList';
-
-export { CalendarAndEventList };
