--- src/pages/dashboard/components/EventList/index.ts
@@ -1,3 +1,0 @@
-import EventList from './EventList';
-
-export { EventList };
