--- src/pages/dashboard/components/MemberList/index.ts
@@ -1,3 +1,0 @@
-import MemberList from './MemberList';
-
-export { MemberList };
