+++ src/pages/dashboard/components/MemberManagement/components/index.ts
@@ -1,0 +1,4 @@
+import MemberList from './MemberList';
+import SearchAndInviteMember from './SearchAndInviteMember';
+
+export { MemberList, SearchAndInviteMember };
