+++ src/pages/dashboard/components/MemberManagement/index.ts
@@ -1,0 +1,3 @@
+import MemberManagement from './MemberManagement';
+
+export { MemberManagement };
