+++ src/pages/dashboard/components/PrescriptionManagement/PrescriptionManagement.tsx
@@ -1,0 +1,5 @@
+const PrescriptionManagement = () => {
+  return <div>PrescriptionManagement</div>;
+};
+
+export default PrescriptionManagement;
