+++ src/pages/dashboard/components/PrescriptionManagement/index.ts
@@ -1,0 +1,3 @@
+import PrescriptionManagement from './PrescriptionManagement';
+
+export { PrescriptionManagement };
