--- src/pages/dashboard/components/SearchAndInviteMember/index.ts
@@ -1,3 +1,0 @@
-import SearchAndInviteMember from './SearchAndInviteMember';
-
-export { SearchAndInviteMember };
