+++ src/pages/dashboard/components/index.ts
@@ -1,19 +1,15 @@
 import { MenuContainer } from './MenuContainer';
 import { Profile } from './Profile';
-import { SearchAndInviteMember } from './SearchAndInviteMember';
-import { MemberList } from './MemberList';
 import { NotSelected } from './NotSelected';
 import { MemberDetail } from './MemberDetail';
-import { EventList } from './EventList';
-import { CalendarAndEventList } from './CalendarAndEventList';
+import { PrescriptionManagement } from './PrescriptionManagement';
+import { MemberManagement } from './MemberManagement';
 
 export {
   MenuContainer,
   Profile,
-  SearchAndInviteMember,
-  MemberList,
   NotSelected,
   MemberDetail,
-  EventList,
-  CalendarAndEventList,
+  PrescriptionManagement,
+  MemberManagement,
 };
